TACO-466: added extra protection for creating a new XioServiceLocator
